third_party/WebKit/LayoutTests/http/tests/priorities/resource-load-priorities-service-worker.https-expected.txt

This is a testharness.js-based test.
PASS registering service worker
PASS Requests from the Fetch API passing through a Service Worker should be loaded with kHigh priority
PASS Render-blocking style sheets requests passing through a Service Worker should be loaded with kVeryHigh priority
FAIL Async scripts passing through a Service Worker should be loaded with kLow priority assert_equals: expected 1 but got 3
FAIL Deferred scripts passing through a Service Worker should be loaded with kLow priority assert_equals: expected 1 but got 3
PASS Module scripts passing through a Service Worker should be loaded with kHigh priority
PASS XHRs passing through a Service Worker should be loaded with kHigh priority
PASS Parser-blocking scripts passing through a Service Worker should be loaded with kHigh priority
FAIL Off-screen images passing through a Service Worker should be loaded with kLow priority assert_equals: expected 1 but got 3
FAIL Prefetches passing through a Service Worker should be loaded with kLowest priority assert_equals: expected 0 but got 3
PASS unregistering service worker
Harness: the test ran to completion.
